EC/CUDA: use multiple streams
