+++ src/components/ec/cuda/ec_cuda_executor_interruptible.c
@@ -1,0 +1,157 @@
+/**
+ * Copyright (C) Mellanox Technologies Ltd. 2020-2022.  ALL RIGHTS RESERVED.
+ *
+ * See file LICENSE for terms.
+ */
+
+#include "ec_cuda_executor.h"
+#include "components/mc/ucc_mc.h"
+
+ucc_status_t ucc_cuda_executor_interruptible_get_stream(cudaStream_t *stream)
+{
+    static int last_used = 0;
+    cudaError_t cuda_st;
+    int i, j;
+
+    ucc_spin_lock(&ucc_ec_cuda.init_spinlock);
+    if (ucc_unlikely(ucc_ec_cuda.exec_streams[0] == NULL)) {
+        for(i = 0; i < EC_CUDA_CONFIG->exec_num_streams; i++) {
+            cuda_st = cudaStreamCreateWithFlags(&ucc_ec_cuda.exec_streams[i],
+                                                cudaStreamNonBlocking);
+            if (ucc_unlikely(cuda_st != cudaSuccess)) {
+                ec_error(&ucc_ec_cuda.super,
+                         "failed to allocate cuda stream %d(%s)",
+                         cuda_st, cudaGetErrorString(cuda_st));
+                for (j = 0; j < i; j++) {
+                    cudaStreamDestroy(ucc_ec_cuda.exec_streams[j]);
+                    ucc_ec_cuda.exec_streams[j] = NULL;
+                }
+                ucc_spin_unlock(&ucc_ec_cuda.init_spinlock);
+                return UCC_ERR_NO_RESOURCE;
+            }
+        }
+    }
+    *stream = ucc_ec_cuda.exec_streams[last_used %
+                                       EC_CUDA_CONFIG->exec_num_streams];
+    last_used++;
+    ucc_spin_unlock(&ucc_ec_cuda.init_spinlock);
+    return UCC_OK;
+}
+
+ucc_status_t
+ucc_cuda_executor_interruptible_task_post(ucc_ee_executor_t *executor,
+                                         const ucc_ee_executor_task_args_t *task_args,
+                                         ucc_ee_executor_task_t **task)
+{
+    ucc_ec_cuda_executor_interruptible_task_t *ee_task;
+    cudaStream_t stream;
+    ucc_status_t status;
+
+    status = ucc_cuda_executor_interruptible_get_stream(&stream);
+    if (ucc_unlikely(status != UCC_OK)) {
+        return status;
+    }
+
+    ee_task = ucc_mpool_get(&ucc_ec_cuda.executor_interruptible_tasks);
+    if (ucc_unlikely(!ee_task)) {
+        return UCC_ERR_NO_MEMORY;
+    }
+
+    status  = ucc_ec_cuda_event_create(&ee_task->event);
+    if (ucc_unlikely(status != UCC_OK)) {
+        ucc_mpool_put(ee_task);
+        return status;
+    }
+
+    ee_task->super.status = UCC_INPROGRESS;
+    ee_task->super.eee    = executor;
+    memcpy(&ee_task->super.args, task_args, sizeof(ucc_ee_executor_task_args_t));
+    switch (task_args->task_type) {
+    case UCC_EE_EXECUTOR_TASK_TYPE_COPY:
+        status = CUDA_FUNC(cudaMemcpyAsync(task_args->bufs[0],
+                                           task_args->bufs[1],
+                                           task_args->count, cudaMemcpyDefault,
+                                           stream));
+        if (ucc_unlikely(status != UCC_OK)) {
+            ec_error(&ucc_ec_cuda.super, "failed to start memcpy op");
+            goto free_task;
+        }
+
+        break;
+    case UCC_EE_EXECUTOR_TASK_TYPE_REDUCE:
+        /* temp workaround to avoid code duplication*/
+        status = ucc_mc_reduce(task_args->bufs[1], task_args->bufs[2],
+                               task_args->bufs[0], task_args->count,
+                               task_args->dt, task_args->op,
+                               UCC_MEMORY_TYPE_CUDA);
+        if (ucc_unlikely(status != UCC_OK)) {
+            ec_error(&ucc_ec_cuda.super, "failed to start reduce op");
+            goto free_task;
+        }
+
+        break;
+    default:
+        ec_error(&ucc_ec_cuda.super, "executor operation is not supported");
+        status = UCC_ERR_INVALID_PARAM;
+        goto free_task;
+    }
+
+    status = ucc_ec_cuda_event_post(stream, ee_task->event);
+    if (ucc_unlikely(status != UCC_OK)) {
+        goto free_task;
+    }
+
+    *task = &ee_task->super;
+    return UCC_OK;
+
+free_task:
+    ucc_ec_cuda_event_destroy(ee_task->event);
+    ucc_mpool_put(ee_task);
+    return status;
+}
+
+ucc_status_t
+ucc_cuda_executor_interruptible_task_test(const ucc_ee_executor_task_t *task)
+{
+    ucc_ec_cuda_executor_interruptible_task_t *ee_task =
+        ucc_derived_of(task, ucc_ec_cuda_executor_interruptible_task_t);
+
+    ee_task->super.status = ucc_ec_cuda_event_test(ee_task->event);
+    return ee_task->super.status;
+}
+
+ucc_status_t
+ucc_cuda_executor_interruptible_task_finalize(ucc_ee_executor_task_t *task)
+{
+    ucc_ec_cuda_executor_interruptible_task_t *ee_task =
+        ucc_derived_of(task, ucc_ec_cuda_executor_interruptible_task_t);
+    ucc_status_t status;
+
+    status = ucc_ec_cuda_event_destroy(ee_task->event);
+    ucc_mpool_put(task);
+    return status;
+}
+
+ucc_status_t ucc_cuda_executor_interruptible_start(ucc_ee_executor_t *executor)
+{
+    ucc_ec_cuda_executor_t *eee = ucc_derived_of(executor,
+                                                 ucc_ec_cuda_executor_t);
+
+    eee->mode  = UCC_EC_CUDA_EXECUTOR_MODE_INTERRUPTIBLE;
+    eee->state = UCC_EC_CUDA_EXECUTOR_STARTED;
+
+    eee->ops.task_post     = ucc_cuda_executor_interruptible_task_post;
+    eee->ops.task_test     = ucc_cuda_executor_interruptible_task_test;
+    eee->ops.task_finalize = ucc_cuda_executor_interruptible_task_finalize;
+
+    return UCC_OK;
+}
+
+ucc_status_t ucc_cuda_executor_interruptible_stop(ucc_ee_executor_t *executor)
+{
+    ucc_ec_cuda_executor_t *eee = ucc_derived_of(executor,
+                                                 ucc_ec_cuda_executor_t);
+
+    eee->state = UCC_EC_CUDA_EXECUTOR_INITIALIZED;
+    return UCC_OK;
+}
